+++ timelinefx.h
@@ -3233,7 +3233,7 @@ TFX_CUSTOM_EMITTER
 
 		void SetUserData(void *data);
 		void *GetUserData();
-		void SetTimeout(unsigned int frames);
+		void SetTimeout(float frames);
 
 		//Override graph functions for use in update_callback
 		//Some of these change the same state and property values, but they're named differently just to make it clearer as to whether you're overriding kEffect or a kEmitter.
